requirements.txt
fastapi
uvicorn
youtube-transcript-api
openai
elevenlabs
python-dotenv
requests
ffmpeg-python